+++ Docs/Chains.md
@@ -24,3 +24,17 @@
 16. `ResponseDataParserNode` - получам `Json` из `Data`
 
 Эта цепочка **НЕ** содерожит кэширования.
+
+`default<I,O>` - классический запрос. Ожидает данные как на вход так и на выход. (описано выше)
+
+`default<Void, Void>` - цепочка не ожидающая данных на вход и на выход (да, и такое бывает)
+
+`default<I, Void>` - цепочка, ожидающая данные на вход, но не возвращающая данные (сервер отвечает пустым телом)
+
+`default<Void, I>` - цепочка, не ожидающая данные на вход, но ожидающая на выход (классический GET-запрос)
+
+`loadData<Void, Data>` - цепочка, которая просто скачивает нужный файл (например качаем статически раздающийся файл)
+
+`loadData<I, Data>` - цепочка, скачивающая файл и передающая на сервер какие-то данные (и так бывает, да)
+
+`default<I, O> where I.DTO.Raw = MultipartModel<[String : Data]>` - цепочка, которая позволяет отправлять multipart-запросы
