+++ NodeKit/Chains/UrlChainsBuilder.swift
@@ -1,11 +1,3 @@
-//
-//  Chains.swift
-//  CoreNetKit
-//
-//  Created by Александр Кравченков on 28/01/2019.
-//  Copyright © 2019 Кравченков Александр. All rights reserved.
-//
-
 import Foundation
 
 /// Реулизует набор цепочек для отправки URL запросов.
@@ -95,6 +87,10 @@ open class UrlChainsBuilder {
         return LoggerNode(next: voidOutput)
     }
 
+    /// Позволяет загрузить бинарные данные (файл) с сервера без отправки какой-то модели на сервер.
+    ///
+    /// - Parameter config: Конфигурация.
+    /// - Returns: Корневой узел цепочки.
     open func loadData(with config: UrlChainConfigModel) -> Node<Void, Data> {
         let loaderParser = DataLoadingResponseProcessor()
         let errorProcessor = ResponseHttpErrorProcessorNode(next: loaderParser)
@@ -116,6 +112,10 @@ open class UrlChainsBuilder {
         return LoggerNode(next: voidInput)
     }
 
+    /// Позволяет загрузить бинарные данные (файл) с сервера.
+    ///
+    /// - Parameter config: Конфигурация.
+    /// - Returns: Корневой узел цепочки.
     open func loadData<Input>(with config: UrlChainConfigModel) -> Node<Input, Data> where Input: DTOEncodable, Input.DTO.Raw == Json {
 
         let loaderParser = DataLoadingResponseProcessor()
@@ -139,7 +139,13 @@ open class UrlChainsBuilder {
         return LoggerNode(next: configNode)
     }
 
-    open func Test<I, O>(with config: UrlChainConfigModel) -> Node<I, O> where O: DTODecodable, O.DTO.Raw == Json, I: DTOEncodable, I.DTO.Raw == MultipartModel<[String : Data]> {
+
+    /// Формирует цепочку для отправки multipaer-запроса.
+    /// Для работы с этой цепочкой в качестве модели необходимо использовать `MultipartModel`
+    ///
+    /// - Parameter config: Конфигурация.
+    /// - Returns: Корневой узел цепочки .
+    open func `default`<I, O>(with config: UrlChainConfigModel) -> Node<I, O> where O: DTODecodable, O.DTO.Raw == Json, I: DTOEncodable, I.DTO.Raw == MultipartModel<[String : Data]> {
 
         let reponseProcessor = self.serviceChain.urlResponseProcessingLayerChain()
