+++ NodeKit/Core/Convertion/DTOConvertible.swift
@@ -1,11 +1,3 @@
-//
-//  DTOConvertible.swift
-//  CoreNetKit
-//
-//  Created by Александр Кравченков on 18/12/2018.
-//  Copyright © 2018 Кравченков Александр. All rights reserved.
-//
-
 import Foundation
 
 /// Композиция протоколов `DTOEncodable` и `DTODecodable`
