+++ NodeKit/Core/Convertion/Extensions/Array+DtoConvertible.swift
@@ -1,11 +1,3 @@
-//
-//  Array+DtoCOnvertible.swift
-//  CoreNetKit
-//
-//  Created by Александр Кравченков on 04/02/2019.
-//  Copyright © 2019 Кравченков Александр. All rights reserved.
-//
-
 import Foundation
 
 /// Позволяетп редставлять массив с элементами `DTODecodable` как `DTODecodable` в случае, если `Raw == Json`
