+++ NodeKit/Core/Convertion/Extensions/Array+RawMappable.swift
@@ -1,11 +1,3 @@
-//
-//  Array_RawMappable.swift
-//  CoreNetKit
-//
-//  Created by Александр Кравченков on 04/02/2019.
-//  Copyright © 2019 Кравченков Александр. All rights reserved.
-//
-
 import Foundation
 
 /// Ошибки для маппинга массивов в/из JSON
