+++ NodeKit/Core/Convertion/RawMappable.swift
@@ -1,11 +1,3 @@
-//
-//  RawMappable.swift
-//  CoreNetKit
-//
-//  Created by Александр Кравченков on 18/12/2018.
-//  Copyright © 2018 Кравченков Александр. All rights reserved.
-//
-
 import Foundation
 
 /// Словарь вида `[String: Any]`
